Some code improvements (#508)
* unit test for union type with null

* Nullable union test

* Nullable union bug fix

* Fix test for old php version

* Targeting PHP minimum version 8.0

* Targeting PHP minimum version 8.0

* Targeting PHP minimum version 8.0

* Run cs fix

* Expanding php 8.0 features usage

* Expanding php 8.0 features usage

* Expanding php 8.0 features usage

* Datetime class name bug fix

* Failing tests fixed

* Bad params with stdClass test removed. It's already throws an error in php 8.0

* Improving code with php 8.0 features

* Improving code with php 8.0 features

* Improving code with php 8.0 features

* Improving code with php 8.0 features

* Improving code with php 8.0 features

* Improving code with php 8.0 features

* Expanding php 8.0 features

* drop test for php <=7.4

* remove some php 7.4 polyfill

* Version 6.1 docs. Drop all examples for php <8.0.

* Failed tests fix

* Improving code with php 8 features

* Improving code with php 8 features

* Improving code with php 8 features

* Improving code with php 8 features

* Improving code with php 8 features

* Improving code with php 8 features

* Improving code with php 8 features

* Remove php 7.4 from CI tests

* Improving code with php 8 features

* Improving code with php 8 features

* Replacing array_merge with built-in spread syntax for lists due to big performance difference.

* Removed types: Container does not match CI server psr extension classes

* Failing CI test fix

* Failing CI test fix

* Fix types for fields descriptors. That caused misunderstanding in methods purpose...

* Failing CI test fix.

* Failing CI test fix.

* Failing CI test fix. Just remove throws tag

* Failing CI test fix. Just remove throws tag

+++ src/Http/HttpCodeDecider.php
@@ -22,13 +22,12 @@ class HttpCodeDecider implements HttpCodeDeciderInterface
         if ($result->data !== null && empty($result->errors)) {
             return 200;
         }
-
         $status = 0;
         // There might be many errors. Let's return the highest code we encounter.
         foreach ($result->errors as $error) {
             $wrappedException = $error->getPrevious();
             if ($wrappedException !== null) {
-                $code = (int) $wrappedException->getCode();
+                $code = $wrappedException->getCode();
                 if ($code < 400 || $code >= 600) {
                     if (! ($wrappedException instanceof ClientAware) || $wrappedException->isClientSafe() !== true) {
                         // The exception code is not a valid HTTP code. Let's ignore it
@@ -43,10 +42,10 @@ class HttpCodeDecider implements HttpCodeDeciderInterface
             } else {
                 $code = 400;
             }
-            $status = max($status, $code);
+            $status = (int) max($status, $code);
         }
 
-        // If exceptions have been thrown and they have not a "HTTP like code", let's throw a 500.
+        // If exceptions have been thrown, and they have not an "HTTP like code", let's throw a 500.
         if ($status < 200) {
             $status = 500;
         }
